+++ lib/rubocop/cop/style/method_call_with_args_parentheses.rb
@@ -156,7 +156,7 @@ module RuboCop
           return if node.implicit_call?
           return if super_call_without_arguments?(node)
           return if camel_case_method_call_without_arguments?(node)
-          return if eligible_for_parentheses_presence?(node)
+          return if legitimate_call_with_parentheses?(node)
 
           add_offense(node, location: node.loc.begin.join(node.loc.end))
         end
@@ -225,7 +225,7 @@ module RuboCop
           node.camel_case_method? && node.arguments.none?
         end
 
-        def eligible_for_parentheses_presence?(node)
+        def legitimate_call_with_parentheses?(node)
           call_in_literals?(node) ||
             call_with_ambiguous_arguments?(node) ||
             call_in_logical_operators?(node) ||
@@ -252,7 +252,7 @@ module RuboCop
             call_as_argument_or_chain?(node) ||
             hash_literal_in_arguments?(node) ||
             node.descendants.any? do |n|
-              splat?(n) || ternary_if?(n) || logical_operator?(n)
+              ambigious_literal?(n) || logical_operator?(n)
             end
         end
 
@@ -285,6 +285,10 @@ module RuboCop
             allowed_chained_call_with_parentheses?(previous)
         end
 
+        def ambigious_literal?(node)
+          splat?(node) || ternary_if?(node) || regexp_slash_literal?(node)
+        end
+
         def splat?(node)
           node.splat_type? || node.kwsplat_type? || node.block_pass_type?
         end
@@ -299,6 +303,10 @@ module RuboCop
 
         def hash_literal?(node)
           node.hash_type? && node.braces?
+        end
+
+        def regexp_slash_literal?(node)
+          node.regexp_type? && node.loc.begin.source == '/'
         end
       end
     end